Automatically replace lints

Automatically replace lints

Usage

fix(path = ".", linters = NULL, exclude_path = NULL, exclude_linters = NULL)

fix_dir(path = ".", linters = NULL)

fix_package(path = ".", linters = NULL)

fix_text(text, linters = NULL, exclude_linters = NULL)

Arguments

path

A valid path to a file or a directory. Relative paths are accepted.

linters

A character vector with the names of the rules to apply. See the entire list of rules with list_linters().

exclude_path

One or several paths that will be ignored from the path selection.

exclude_linters

One or several linters that will not be checked. Values can be the names of linters (such as "any_is_na") or its associated function, such as any_is_na_linter() (this is mostly for compatibility with lintr).

text

Text to analyze (and to fix if necessary).

Ignoring lines

flint supports ignoring single lines of code with # flint-ignore. For example, this will not warn:

# flint-ignore
any(duplicated(x))

However, this will warn for the second any(duplicated()):

# flint-ignore
any(duplicated(x))
any(duplicated(y))

To ignore more than one line of code, use # flint-ignore-start and # flint-ignore-end:

# flint-ignore-start
any(duplicated(x))
any(duplicated(y))
# flint-ignore-end