List all lints in a file or a directory

lint(), lint_text() and lint_diff() all produce a list of lints. The only difference is in the input they take:

lint() takes path to files or directories

lint_text() takes some text input

lint_diff() takes a path to a directory but only looks at files that have changed since the last commit.

Usage

lint(path = ".", linters = NULL, open = TRUE)

lint_dir(path = ".", linters = NULL, open = TRUE)

lint_package(path = ".", linters = NULL, open = TRUE)

lint_text(text, linters = NULL)

Arguments

path

A valid path to a file or a directory. Relative paths are accepted.

linters

A character vector with the names of the rules to apply. See the entire list of rules with list_linters().

open

If TRUE (default) and if this is used in the RStudio IDE, lints will be shown with markers.

Value

A dataframe where each row is a lint. The columns show the text, its location (both the position in the text and the file in which it was found) and the severity.