List all lints in a file or a directory

lint(), lint_text() and lint_dir() all produce a data.frame containing the lints, their location, and potential fixes. The only difference is in the input they take:

lint() takes path to files or directories

lint_text() takes some text input

lint_dir() takes a path to one directory

lint_package() takes a path to the root of a package and looks at the following list of folders: R, tests, inst, vignettes, data-raw, demo, exec.

Usage

lint(
  path = ".",
  linters = NULL,
  exclude_path = NULL,
  exclude_linters = NULL,
  open = TRUE,
  use_cache = TRUE
)

lint_dir(
  path = ".",
  linters = NULL,
  open = TRUE,
  exclude_path = NULL,
  exclude_linters = NULL,
  use_cache = TRUE
)

lint_package(
  path = ".",
  linters = NULL,
  open = TRUE,
  exclude_path = NULL,
  exclude_linters = NULL,
  use_cache = TRUE
)

lint_text(text, linters = NULL, exclude_linters = NULL)

Arguments

path

A valid path to a file or a directory. Relative paths are accepted.

linters

A character vector with the names of the rules to apply. See the entire list of rules with list_linters().

exclude_path

One or several paths that will be ignored from the path selection.

exclude_linters

One or several linters that will not be checked. Values can be the names of linters (such as "any_is_na") or its associated function, such as any_is_na_linter() (this is mostly for compatibility with lintr).

open

If TRUE (default) and if this is used in the RStudio IDE, lints will be shown with markers.

use_cache

Do not re-parse files that haven't changed since the last time this function ran.

text

Text to analyze.

Value

A dataframe where each row is a lint. The columns show the text, its location (both the position in the text and the file in which it was found) and the severity.

Ignoring lines

flint supports ignoring single lines of code with # flint-ignore. For example, this will not warn:

# flint-ignore
any(duplicated(x))

However, this will warn for the second any(duplicated()):

# flint-ignore
any(duplicated(x))
any(duplicated(y))

To ignore more than one line of code, use # flint-ignore-start and # flint-ignore-end:

# flint-ignore-start
any(duplicated(x))
any(duplicated(y))
# flint-ignore-end

Examples

# `lint_text()` is convenient to explore with a small example
lint_text("any(duplicated(rnorm(5)))")
#> Original code: any(duplicated(rnorm(5))) 
#> Suggestion: anyDuplicated(x, ...) > 0 is better than any(duplicated(x), ...). 

lint_text("any(duplicated(rnorm(5)))
any(is.na(x))
")
#> Original code: any(duplicated(rnorm(5))) 
#> Suggestion: anyDuplicated(x, ...) > 0 is better than any(duplicated(x), ...). 
#> Original code: any(is.na(x)) 
#> Suggestion: anyNA(x) is better than any(is.na(x)). 

# Setup for the example with `lint()`
destfile <- tempfile()
cat("
x = c(1, 2, 3)
any(duplicated(x), na.rm = TRUE)

any(duplicated(x))

if (any(is.na(x))) {
  TRUE
}

any(
  duplicated(x)
)", file = destfile)

lint(destfile)
#>                  id                             text line_start col_start
#>              <char>                           <char>      <num>     <num>
#> 1: equal_assignment                   x = c(1, 2, 3)          2         1
#> 2:   any_duplicated any(duplicated(x), na.rm = TRUE)          3         1
#> 3:   any_duplicated               any(duplicated(x))          5         1
#> 4:           any_na                    any(is.na(x))          7         5
#> 5:   any_duplicated         any(\n  duplicated(x)\n)         11         1
#>    line_end col_end severity                          fix
#>       <num>   <num>   <char>                       <char>
#> 1:        2      15     hint           ~~LHS~~ <- ~~RHS~~
#> 2:        3      33  warning anyDuplicated(~~MYVAR~~) > 0
#> 3:        5      19  warning anyDuplicated(~~MYVAR~~) > 0
#> 4:        7      18  warning             anyNA(~~MYVAR~~)
#> 5:       13       2  warning anyDuplicated(~~MYVAR~~) > 0
#>                                                              message
#>                                                               <char>
#> 1:                                    Use <-, not =, for assignment.
#> 2: anyDuplicated(x, ...) > 0 is better than any(duplicated(x), ...).
#> 3: anyDuplicated(x, ...) > 0 is better than any(duplicated(x), ...).
#> 4:                            anyNA(x) is better than any(is.na(x)).
#> 5: anyDuplicated(x, ...) > 0 is better than any(duplicated(x), ...).
#>                                file
#>                              <char>
#> 1: /tmp/RtmpP0G2PA/file179b43bbada4
#> 2: /tmp/RtmpP0G2PA/file179b43bbada4
#> 3: /tmp/RtmpP0G2PA/file179b43bbada4
#> 4: /tmp/RtmpP0G2PA/file179b43bbada4
#> 5: /tmp/RtmpP0G2PA/file179b43bbada4